wii: add smaller defconfig
This is a smaller defconfig for the Wii.
It will be used to build a smaller
recovery / loader environment, known as
loader.img.  Please do not attempt to
use this defconfig as-is to build just
a smaller kernel.  You will be missing
a lot of functionality.  This is designed
to be a smaller set of modules for the
fully-featured kernel.