ehci: disable static functions & wii fixes
Remove "static" from the function definitions,
this makes it lose performance, but makes it easier
to debug when they are failing, as they are now
included in the kernel stack trace.

Also, force debug printing during submit_async,
it's the only thing I've found that actually fixes
the panics caused in it on the Wii.